# server/contracts/openapi/homeos.yaml
...
openapi: "3.0.3"
info:
  title: HomeOS API
  version: "1.0"
  description: >
    HomeOS 底座服务接口，负责身份、家庭、权限、通知、同步、搜索等基础能力。
    所有接口均需携带 JWT token（除健康检查与 JWKS 分发外）。
servers:
  - url: /api/homeos
paths:
...
  /family/modules:
    get:
      summary: 获取家庭面配置与全量面目录
      description: >
        每项含 code/name/icon/enabled（家庭口径）/visible（当前成员角色口径）/born/availability。
        开通页直接读它；首页矩阵渲染数据取自 home/summary 的 faces[]。
      operationId: getFamilyModules
      tags:
        - family
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 面配置列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  modules:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        name:
                          type: string
                        icon:
                          type: string
                        enabled:
                          type: boolean
                        visible:
                          type: boolean
                          description: 当前成员角色是否可见
                        born:
                          type: boolean
                          description: 服务是否已出生
                        availability:
                          type: string
                          enum: [available, unavailable]
                  version:
                    type: integer

    put:
      summary: 管理员开通/停用某个面
      description: >
        仅管理员可调用。落审计、pver+1、发 homeos.family.module.updated 事件。
        开通只写配置，不创建该面任何业务数据；停用不删数据。
        仅剩最后一个已启用面时停用被拒。
      operationId: updateFamilyModules
      tags:
        - family
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, enabled, version]
              properties:
                code:
                  type: string
                enabled:
                  type: boolean
                version:
                  type: integer
                  description: 乐观锁版本号
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: integer
                  pver:
                    type: integer
        '403':
          description: 非管理员无权调用
        '409':
          description: 版本冲突或仅剩最后一面不允许停用

  # === 邀请与加入家庭（PRD 3.4.1 接口表：POST /members/invite、POST /family/invite/accept、
  #     DELETE /family/invites/{id}；GET /family/invites 为前端在册的邀请名册读口，
  #     PRD 接口表未登记该读口——已作为定版空白上报，待人工裁定是否补入 3.4.1） ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT